Refactor stocktake delete hook
