+++ client/packages/inventory/src/Stocktake/api/hooks/document/index.ts
@@ -1,6 +1,5 @@
 import { useStocktake } from './useStocktake';
 import { useStocktakeDelete } from './useStocktakeDelete';
-import { useStocktakeDeleteSelected } from './useStocktakeDeleteSelected';
 import { useStocktakeFields } from './useStocktakeFields';
 import { useStocktakes } from './useStocktakes';
 import { useStocktakesAll } from './useStocktakesAll';
@@ -9,7 +8,6 @@ import { useUpdateStocktake } from './useUpdateStocktake';
 export const Document = {
   useStocktake,
   useStocktakeDelete,
-  useStocktakeDeleteSelected,
   useStocktakeFields,
   useStocktakes,
   useStocktakesAll,
